pyproject.toml
[project]
name = "compound-ai-systems"
version = "0.1.0"
description = "Compound AI Systems Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "instructor>=1.7.9",
    "isort>=6.0.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "python-decouple==3.8",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "uvicorn>=0.34.2",
]
